Add files via upload
I'm create a website using a combination of technologies including HTML, CSS, JavaScript, React, Node.js, and MongoDB. HTML is used for creating the structure and layout of the website. CSS is used for styling and visual design. JavaScript is used for adding interactivity and dynamic functionality to the website. React is a JavaScript library used for building user interfaces. Node.js is a JavaScript runtime used for building server-side applications. MongoDB is a NoSQL database used for storing and retrieving data for the website.
And it is also as a client side based.